requirements.txt
Django<=1.10
psycopg2
pyelasticsearch
django-appconf
pyshp
dj-database-url
opencivicdata-divisions
DjangoRestless

-e git://github.com/opencivicdata/pupa.git#egg=pupa
-e git://github.com/opencivicdata/imago.git#egg=imago

represent-boundaries
us
